+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # CHANGELOG
 
+## v7.0.1 (2018-05-04)
+- Check for undefined values on pack().
+- Check for null values on unpack().
+
 ## v7.0.0 (2018-05-03)
 - Type class is deprecated. Types should be defined as Object<string, *>.